+++ test/bigtest/interactors/textarea-interactor.js
@@ -1,0 +1,22 @@
+import {
+  blurrable,
+  fillable,
+  focusable,
+  interactor,
+  value,
+} from '@bigtest/interactor';
+
+@interactor
+export class TextAreaInteractor {
+  val = value();
+  fillTextArea = fillable();
+  focusTextArea = focusable();
+  blurTextArea = blurrable();
+
+  fillAndBlur(val) {
+    return this
+      .focusTextArea()
+      .fillTextArea(val)
+      .blurTextArea();
+  }
+}