+++ test/bigtest/tests/match-profiles/new-match-profiles-test.js
@@ -44,7 +44,7 @@ describe('Match profile form', () => {
       this.visit('/settings/data-import/match-profiles?layer=create');
     });
 
-    it('when not filled when the submit button is disabled', () => {
+    it('when not filled then the submit button is disabled', () => {
       expect(matchProfileForm.submitFormButtonDisabled).to.be.true;
     });
 
@@ -60,26 +60,68 @@ describe('Match profile form', () => {
     });
 
     describe('details accordion', () => {
-      it('show initial record select view', () => {
-        expect(matchProfileForm.recordTypesSelect.initialRecord).to.be.true;
-      });
-
-      it('incoming record select has correct amount of items', () => {
-        expect(matchProfileForm.recordTypesSelect.items().length).to.be.equal(8);
-      });
-
-      describe('when incoming record is selected', () => {
-        beforeEach(async () => {
-          await matchProfileForm.recordTypesSelect.select('ITEM');
-        });
-
-        it('show compare record select view', () => {
-          expect(matchProfileForm.recordTypesSelect.initialRecord).to.be.false;
-          expect(matchProfileForm.recordTypesSelect.compareRecord).to.be.true;
-        });
-
-        it('then choose record to compare screen appears', () => {
+      describe('"Record types select" component', () => {
+        it('should render', () => {
           expect(matchProfileForm.recordTypesSelect.isPresent).to.be.true;
+        });
+
+        it('should show initial record select view', () => {
+          expect(matchProfileForm.recordTypesSelect.initialRecord).to.be.true;
+        });
+
+        it('incoming record select has correct amount of items', () => {
+          expect(matchProfileForm.recordTypesSelect.items().length).to.be.equal(8);
+        });
+
+        describe('when incoming record is selected', () => {
+          beforeEach(async () => {
+            await matchProfileForm.recordTypesSelect.select('ITEM');
+          });
+
+          it('should show compare record select view', () => {
+            expect(matchProfileForm.recordTypesSelect.initialRecord).to.be.false;
+            expect(matchProfileForm.recordTypesSelect.compareRecord).to.be.true;
+          });
+
+          it('then choose record to compare screen appears', () => {
+            expect(matchProfileForm.recordTypesSelect.isPresent).to.be.true;
+          });
+        });
+      });
+
+      describe('"Match criterion" component', () => {
+        it('should render', () => {
+          expect(matchProfileForm.matchCriteria.isPresent).to.be.true;
+        });
+
+        describe('"MARC field in incoming record" section', () => {
+          describe('when not filled', () => {
+            it('should have empty "Field" input', () => {
+              expect(matchProfileForm.matchCriteria.inputMain.val).to.be.equal('');
+            });
+
+            it('should have empty "In.1" input', () => {
+              expect(matchProfileForm.matchCriteria.inputIn1.val).to.be.equal('');
+            });
+
+            it('should have empty "In.2" input', () => {
+              expect(matchProfileForm.matchCriteria.inputIn2.val).to.be.equal('');
+            });
+
+            it('should have empty "Subfield" input', () => {
+              expect(matchProfileForm.matchCriteria.inputSubfield.val).to.be.equal('');
+            });
+          });
+        });
+
+        describe('"Match criterion" section', () => {
+          it('should render', () => {
+            expect(matchProfileForm.matchCriteria.matchCriterionSection.isPresent).to.be.true;
+          });
+
+          it('should have a dropdown', () => {
+            expect(matchProfileForm.matchCriteria.matchCriterionSection.dropdown.isPresent).to.be.true;
+          });
         });
       });
     });